refactor(editor): table width 모델을 분리한다
